refactor auth context and registration
